Simulations MV Screening
R code used to produce the simulations found in Ultrahigh Dimensional Screening with Multivariate Response. User will need to comment and uncomment the coefficient generations for each of parts A, B, C in Sims 1 and 3. In final paper, Sim 5 is 5.B, Sim 6 is 5.C and Sim 7 is 5.A. 
Choice of norm (1 or 2) will also need to be adjusted as desired.